Move test remaining database helper functions into database file, remove un-needed and unused monkeypatch code in mountain project api tests.
